backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
supabase==2.3.4
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
pydantic-settings==2.1.0

# Data processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.11.0
pypdf==4.0.1

# LLM Integration (OpenRouter)
openai==1.12.0
httpx==0.26.0
aiohttp==3.9.3

# Banking APIs
plaid-python==20.0.0

# Additional utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic[email]==2.6.1

# Caching (optional)
redis==5.0.1

# Image processing (for PDF with images)
Pillow==10.2.0

# Data validation
email-validator==2.1.0.post1

# Date handling
python-dateutil==2.8.2

# HTTP requests
requests==2.31.0

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.4

# Logging
structlog==24.1.0
